Support both assets (wla://) and jar files (wlj://) in the asset manager. JARs that are contained within modules have the form wlj://<module name>/<plugin name>/<path> where <path> is something like "common/my_jar_common.jar". Use the normal URL.openStream() mechanism to fetch assets of this type.
The ModuleUtils.fetchPluginJars() returns a ModulePluginList object of all plugins
needed on the client. Useful code for fetching all of the needed JARs is:

        ModulePluginList list = ModuleUtils.fetchPluginJars();
        if (list == null) {
            System.out.println("NULL PLUGIN LIST");
            return;
        }
        for (String uri : list.getJarURIs()) {
            System.out.println("PLUGIN " + uri);
            try {
                URL url = new URL(uri);
                url.openStream();
            } catch (Exception excp) {
                excp.printStackTrace();
           }
        }

Note that checksums need to be generated for plugin jars now too (this is done
automatically during module installation).

Also updated how the asset manager stores assets in the cache. In the cache,
assets belonging to a module are found under <module name>/ directory. In
that, art/ and plugins/ cache art assets and plugin jars, respectively. The
code now also implements the spec where the name of the asset is a directory,
beneath are the actual files of the assets, named by their checksum. (This
lets multiple assets with the same name be cached).

Check in initial (untested) code to launch an external browser cross-
platform.



git-svn-id: file:///home/wonderland/wl-googlecode-mirror/openwonderland-mirror/trunk@461 f741164b-69bb-45d8-84fe-596f280cac40