Added some mechanism to create and delete cells.
1. Create a mechanism by which modules can register their cells to handle media with a
certain file extension. Modules must implement the CellExtensionTypeSPI and register
their class in META-INF/services. The module build ant task supports this with the
<cellextensiontype> tag (within <serverplugin>), similar to how the <cellsetup> tag
is supported. The idea is that the class returns a new cell setup class that the
server can create given the media asset uri (and its file extension). We need this
sort of mechanism because the Darkstar server must know how to create a new cell,
and all it has is perhaps the uri of an asset (in a module). Once the DS server
has the setup class, it can create a new cell.

2. Create a communication mechanism to create and delete cells. For now, it assumes
the cell is based upon some media asset uri. (e.g. COLLADA models). The server-side
communication channel is registered on CelllManagerMO (and the client-side handler
in CellClientSession). There are two messages supported: CellCreateMessage and
CellDeleteMessage.

For example, to create a new cell:

    WonderlandSession session = LoginManager.getPrimary().getPrimarySession();
    CellEditChannelConnection connection = (CellEditChannelConnection)session.getConnection(CellEditConnectionType.CLIENT_TYPE);
    String assetURI = "wla://" + moduleName + "/" + artPath;
    CellID parentCellID = ....;
    CellCreateMessage msg = new CellCreateMessage(parentCellID, assetURI);
    connection.send(msg); 

3. Oh, also cleaned up some log messages there were mistakening WARNING.



git-svn-id: file:///home/wonderland/wl-googlecode-mirror/openwonderland-mirror/trunk@970 f741164b-69bb-45d8-84fe-596f280cac40