Major refactor of the client-side asset manager. The motivating factor was to support both checksum-based assets and versioning done by the HTTP if-modified-since header. There have been no changes to the public API (e.g. AssetManager.getAsset()), the changes should be entirely internal. Also implemented an Asset Meter, a window that tracks the progress of assets being downloaded and cached by the Asset Manager
The Asset Manager gets a list of repositories from where to download an asset via an AssetRepositoryFactory
class. This generates a list of AssetRepository classes, that in turn returns AssetStream classes that
indicates the asset is already cache or provides an input stream to download the asset. There is an
implementation of these three classes for each mechanism: checksum-based assets and if-modified-since
based assets.



git-svn-id: file:///home/wonderland/wl-googlecode-mirror/openwonderland-mirror/trunk@1921 f741164b-69bb-45d8-84fe-596f280cac40