Remove Call when call ends on its own.
Remove private spatializer when volume is 1 and the spatializer is no longer needed.

Don't disply virtual orb's in user lists.

Don't put up an orb dialog for a virtual orb of yourself.

Fix problem where adding a user to an audio group caused the chat type to change
to private for the person adding the user.

Put virtual orb in correct initial position.

Add action listeners to InCall privacy radio buttons.

Don't crash if user tries to place a call and the softphon isn't connected.
Instead display a message in the PlaceCallDialog.




git-svn-id: file:///home/wonderland/wl-googlecode-mirror/openwonderland-mirror/trunk@2334 f741164b-69bb-45d8-84fe-596f280cac40