Fix bug #383: Fix multi-threading issue in listener notification
There are no developer visible changes in this bug fix.

The event notification method (notifyEventListeners) in HUDObject2D was not re-entrant.
It was possible for this method to be invoked multiple times while still delivering
events. This resulted in corruption of the previous event delivery, resulting in lost
events. The most obvious example of this was the stuck "loading avatar..." HUD message.
This method now uses a thread-safe event queue which ensures that events are not lost
and are delivered in the order in which they occurred.

I also simplified the event listening design. Event sources such as HUDComponents
notify their containers (HUD) which in turn notify their listeners (e.g.,
HUDComponentManagers). Previously HUDComponentManagers listened to both HUDs and
HUDComponents which lead to events being delivered out of order. This resulted in
situations where a component would be added to a HUD and made visible, but the 
visible event would be delivered before the added event (out of time order) resulting
in the component never appearing.

I'm suspicious of there being increased deadlock potential in this fix, so I'm going
to leave the bug open until we've done more testing.



git-svn-id: file:///home/wonderland/wl-googlecode-mirror/openwonderland-mirror/trunk@3135 f741164b-69bb-45d8-84fe-596f280cac40